+++ src/sorting/algorithms/selection-sort.js
@@ -1,13 +1,25 @@
-async function selectionSort(arr, swap, delay){
+import { swap } from "../utils/array-utils";
+import sleep from "../utils/sleep";
+
+
+/**
+ * The implementation of selection sort
+ * @param {Array<Number>} arr array
+ * @param {CallableFunction} updateBar callback function
+ * @param {Number} delay delay in ms
+ */
+async function selectionSort(arr, updateBar, delay){
     for(let i=0; i<arr.length; i++){
         let minIndex = i;
         for(let j=i+1; j<arr.length; j++){
             if (arr[j] < arr[minIndex]){
-                minIndex = j;   
+                minIndex = j;  
+                //array accessed two times in if statement
             };
         };
-        await swap(arr, minIndex, i, delay);
-        
+        swap(arr, minIndex, i);
+        await updateBar(arr, minIndex, delay);
+        await updateBar(arr, i, delay);
     };
 };